Added dice command.
